+++ flumine/markets/blotter.py
@@ -197,13 +197,20 @@ class Blotter:
                 _size_matched = order.size_matched  # cache
                 _order_side = order.side
                 if _size_matched:
-                    if _order_side == "BACK":
-                        mb.append((order.average_price_matched, _size_matched))
+                    if order.order_type.price_ladder_definition == "LINE_RANGE":
+                        average_price_matched = 2.0
                     else:
-                        ml.append((order.average_price_matched, _size_matched))
+                        average_price_matched = order.average_price_matched
+                    if _order_side == "BACK":
+                        mb.append((average_price_matched, _size_matched))
+                    else:
+                        ml.append((average_price_matched, _size_matched))
                 if not order.complete:
                     _size_remaining = order.size_remaining  # cache
-                    order_type_price = order.order_type.price
+                    if order.order_type.price_ladder_definition == "LINE_RANGE":
+                        order_type_price = 2.0
+                    else:
+                        order_type_price = order.order_type.price
                     if order_type_price and _size_remaining:
                         if _order_side == "BACK":
                             ub.append((order_type_price, _size_remaining))