+++ flumine/order/order.py
@@ -75,6 +75,7 @@ class BaseOrder:
 
         self.client = None
         self.runner_status = None  # RunnerBook.status
+        self.line_range_result = None
         self.market_type = None
         self.each_way_divisor = 1
         self.number_of_dead_heat_winners = None
@@ -303,6 +304,7 @@ class BaseOrder:
                 "elapsed_seconds_executable": self.elapsed_seconds_executable,
             },
             "runner_status": self.runner_status,
+            "line_range_result": self.line_range_result,
             "status": self.status.value if self.status else None,
             "status_log": ", ".join([s.value for s in self.status_log]),
             "violation_msg": self.violation_msg,