+++ HISTORY.rst
@@ -11,6 +11,10 @@ Release History
 - #725 Remove redundant event logging
 - Refactor _process_sports_data to use eventId
 
+**Bug Fixes**
+
+- Correctly handle LINE_RANGE markets within trading controls / blotter / profit / order
+
 2.5.4 (2023-11-13)
 +++++++++++++++++++